refactor(webui): isolate navigation shell
